pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datashield"
version = "0.1.0"
description = "The missing validation layer between unstructured data and AI systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "DataShield Team", email = "team@datashield.dev" }
]
keywords = [
    "data-validation",
    "ai",
    "llm",
    "rag",
    "multimodal",
    "compliance",
    "hipaa",
    "gdpr",
    "langchain",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "orjson>=3.9",
    "pandas>=2.0",
    "pdfplumber>=0.10",
    "jsonschema>=4.0",
    "litellm>=1.0",
    "langchain-core>=0.1",
    "python-dateutil>=2.8",
    "rich>=13.0",
]

[project.optional-dependencies]
ocr = ["pytesseract", "pillow"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "pandas-stubs>=2.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]
all = ["datashield[ocr,dev,docs]"]

[project.scripts]
datashield = "datashield.cli.main:cli"

[project.urls]
Homepage = "https://github.com/datashield/datashield"
Documentation = "https://datashield.dev/docs"
Repository = "https://github.com/datashield/datashield"
Issues = "https://github.com/datashield/datashield/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/datashield"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["datashield"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "pdfplumber.*",
    "litellm.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q --cov=datashield --cov-report=term-missing"

[tool.coverage.run]
source = ["src/datashield"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
